Hilbert transform and codeGen

+++ include/LSModulation.h
@@ -39,7 +39,7 @@ double code(double now);
 double step(double now);
 
 ModParameters::modF getSinMod(double time, int number);
-ModParameters::modF getCodeMod(double discret, std::vector<int> code);
+ModParameters::modF getCodeMod(double discret, std::vector<double> code);
 
 
 /*===================================================================================================================*/
